+++ ChangeLog.md
@@ -2,3 +2,7 @@
 - created change log
 - created basic dev work
 - started work on basic lootboxes
+- created basic shop
+
+5/30/2024
+- created lootbox class